Refactor authenticationHelpers into service directory format
